+++ NEWS.md
@@ -1,3 +1,8 @@
+# unjoin dev
+
+* fix bug that was preventing key_col arg from working
+
+
 # unjoin 0.0.0.9000
 
 * first working build